… include/claim_strategy.hpp
… include/tag.hpp
… test/qa_buffer.cpp
… test/qa_scheduler.cpp
… test/qa_settings.cpp